Revert "Feat/record diet"
