--- fitty_frontend_app/lib/class/food_name_and_amount.dart
@@ -1,6 +1,0 @@
-class FoodNameAndAmount {
-  String name;
-  int amount;
-
-  FoodNameAndAmount(this.name, this.amount);
-}
