+++ fitty_frontend_app/lib/page/calendar_menu/tab/my_exercise_calendar_tab.dart
@@ -1,4 +1,4 @@
-import 'package:fitty_frontend_app/data/my_exercise_record.dart';
+import 'package:fitty_frontend_app/data/all_exercise_record.dart';
 import 'package:fitty_frontend_app/page/exercise/daily_routine_page.dart';
 import 'package:fitty_frontend_app/widget/exercise/one_exercise_record_summary_widget.dart';
 import 'package:fitty_frontend_app/widget/exercise/time_record_widget.dart';
@@ -43,7 +43,7 @@ class _MyExerciseCalendarTabState extends State<MyExerciseCalendarTab> {
                   fontWeight: FontWeight.bold,
                 )),
           ),
-          Consumer<MyExerciseRecord>(
+          Consumer<AllExerciseRecord>(
             builder: (context, allExerciseRecord, child) {
               var selectedDayExerciseRecord =
                   allExerciseRecord.getDayExerciseRecord(_selectedDay);
