+++ fitty_frontend_app/lib/page/exercise/daily_routine_page.dart
@@ -1,11 +1,9 @@
-import 'package:fitty_frontend_app/data/my_exercise_record.dart';
-import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
+import 'package:fitty_frontend_app/data/all_exercise_record.dart';
 import 'package:fitty_frontend_app/page/exercise/exercise_record_list_page.dart';
 import 'package:fitty_frontend_app/widget/exercise/one_exercise_record_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:provider/provider.dart';
-import '../../data/class/day_exercise_record.dart';
 import 'exercise_list_page.dart';
 
 class DailyRoutinePage extends StatelessWidget {
@@ -46,7 +44,7 @@ class DailyRoutinePage extends StatelessWidget {
           title: Text(title),
         ),
         body: SingleChildScrollView(
-          child: Center(child: Consumer<MyExerciseRecord>(
+          child: Center(child: Consumer<AllExerciseRecord>(
             builder: (context, allExerciseRecord, child) {
               var selectedDayExerciseRecord =
                   allExerciseRecord.getDayExerciseRecord(selectedDay);
