--- fitty_frontend_app/lib/page/friend_menu/friend_page.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-class FriendPage extends StatelessWidget {
-  const FriendPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    // todo : 친구 목록을 불러와서 보여주는 페이지, 출시 이후 구현 예정
-    return Container();
-  }
-}
