+++ fitty_frontend_app/lib/widget/exercise/one_exercise_record_summary_widget.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
-import 'package:fitty_frontend_app/data/my_exercise_record.dart';
-import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
+import 'package:fitty_frontend_app/data/all_exercise_record.dart';
 import 'package:flutter/material.dart';
 
 class OneExerciseRecordSummaryWidget extends StatelessWidget {
