+++ fitty_frontend_app/linux/.gitignore
