+++ fitty_frontend_app/linux/CMakeLists.txt
