+++ fitty_frontend_app/linux/flutter/CMakeLists.txt
