+++ fitty_frontend_app/linux/flutter/generated_plugins.cmake
