+++ fitty_frontend_app/linux/my_application.h
